testGUI is now start.py
