+++ start.py
@@ -2,213 +2,213 @@
 
 import os
 import csv
-import web
 import sys
-import time
 import socket
 import threading
-import webbrowser
-import subprocess
 
-from web import form
-from time import sleep
-import youtubeup as yup
-from datetime import datetime
-from argparse import Namespace
 import youtubeAuthenticate as YA
-
-from consts import *
-
-render = web.template.render('webpage/')
-
-dataform = form.Form(
-    form.Dropdown("where",
-                  [("../", "Parent Folder to Scripts"),
-                   ("", "Same Folder as Scripts")],
-                  description="Match Files Location"),
-    form.Textbox("prodteam", description="Production Team", size=41),
-    form.Textbox("twit",
-                 form.Validator("Remove @", lambda x: x[0] != "@"),
-                 description="Twitter Handle", size=41),
-    form.Textbox("fb", description="Facebook Name", size=41),
-    form.Textbox("weblink", description="Website Link", size=41),
-    form.Textbox("ename", description="Event Name", size=41),
-    form.Textbox("ecode", description="Event Code (ex. 2016arc)"),
-    form.Textbox("pID",
-                 form.regexp(
-                     "^PL", "Must be a playlist ID, all of which start with 'PL'"),
-                 form.regexp("^\s*\S+\s*$", "Can not contain spaces."),
-                 description="Playlist ID",
-                 size=41),
-    form.Textbox("tbaID",
-                 form.Validator("Make sure there are no spaces", lambda x: " " not in x or x ==
-                                "Go to thebluealliance.com/request/apiwrite to get keys"),
-                 description="TBA Event ID",
-                 value="Go to thebluealliance.com/request/apiwrite to get keys",
-                 size=41),
-    form.Textbox("tbaSecret",
-                 form.Validator("Make sure there are no spaces", lambda x: " " not in x or x ==
-                                "Go to thebluealliance.com/request/apiwrite to get keys"),
-                 description="TBA Event Secret",
-                 value="Go to thebluealliance.com/request/apiwrite to get keys",
-                 size=41),
-    form.Textarea("description",
-                  description="Video description",
-                  value="Add alternate description here."),
-    form.Textbox("mcode",
-                 form.Validator("Must be 0 if not used properly", lambda x: any(
-                     k in x for k in ("qm", "qf", "sf", "f1m", "0")) or x == ""),
-                 value="0",
-                 description="Match Code"),
-    form.Textbox("mnum",
-                 form.notnull,
-                 form.regexp("\d+", "Cannot contain letters"),
-                 form.Validator("Must be more than 0", lambda x: int(x) > 0),
-                 description="Match Number"),
-    form.Dropdown("mtype",
-                  [("qm", "Qualifications"), ("qf", "Quarterfinals"),
-                   ("sf", "Semifinals"), ("f1m", "Finals")],
-                  description="Match Type"),
-    form.Dropdown("tiebreak", [("no", "False"),
-                               ("yes", "True")], description="Tiebreaker"),
-    form.Dropdown("tba", [("yes", "True"), ("no", "False")],
-                  description="Use TBA"),
-    form.Dropdown("ceremonies", [(0, "None"), (1, "Opening Ceremonies"), (
-        2, "Alliance Selection"), (3, "Closing Ceremonies")], description="Ceremonies"),
-    form.Dropdown("eday", [(0, "Ignore"), (1, "1"),
-                           (2, "2"), (3, "3")], description="Event Day"),
-    form.Textbox("end",
-                 description="Last Match Number",
-                 value="Only for batch uploads"),
-    validators=[form.Validator("Last Match Number must be greater than Match Number",
-                               lambda i: i.end == "Only for batch uploads" or int(i.end) > int(i.mnum))]
-)
+import youtubeup as yup
 
 
-class index(threading.Thread):
-    def run(self):
-        urls = ('/', 'index')
-        app = web.application(urls, globals())
-        if DEBUG:
-            web.httpserver.runsimple(app.wsgifunc(), ("0.0.0.0", 8888))
-            webbrowser.open_new("http://localhost:8888")
-        else:
-            webbrowser.open_new("http://localhost:8080")
-        app.run()
+from PyQt5 import QtCore
+from datetime import datetime
+import pyforms
+from argparse import Namespace
+from pyforms import BaseWidget
+from pyforms.controls import ControlText
+from pyforms.controls import ControlTextArea
+from pyforms.controls import ControlCombo
+from pyforms.controls import ControlButton, ControlCheckBox
 
-    def GET(self):
-        myform = dataform()
+
+class EmittingStream(QtCore.QObject):
+
+    textWritten = QtCore.pyqtSignal(str)
+
+    def write(self, text):
+        self.textWritten.emit(str(text))
+
+    def flush(self):
+        pass
+
+
+class FRC_Uploader(BaseWidget):
+
+    def __init__(self):
+        super(FRC_Uploader, self).__init__("FRC YouTube Uploader")
+        # Redirct print output
+        sys.stdout = EmittingStream(textWritten=self.writePrint)
+        # Create form fields
+        self._topbutton = ControlButton('Submit')
+        # Event Values
+        self._where = ControlCombo("Match Files Location")
+        self._prodteam = ControlText("Production Team")
+        self._twit = ControlText("Twitter Handle")
+        self._fb = ControlText("Facebook Name")
+        self._weblink = ControlText("Website Link")
+        self._ename = ControlText("Event Name")
+        self._ecode = ControlText("Event Code")
+        self._pID = ControlText("Playlist ID")
+        self._tbaID = ControlText("TBA ID")
+        self._tbaSecret = ControlText("TBA Secret")
+        self._description = ControlTextArea("Video Description")
+        # Match Values
+        self._mcode = ControlText("Match Code")
+        self._mnum = ControlText("Match Number")
+        self._mtype = ControlCombo("Match Type")
+        self._tiebreak = ControlCheckBox("Tiebreaker")
+        self._tba = ControlCheckBox("Use TBA")
+        self._ceremonies = ControlCombo("Ceremonies")
+        self._eday = ControlCombo("Event Day")
+        self._end = ControlText("Last Match Number")
+
+        # Output Box
+        self._output = ControlTextArea()
+        self._output.readonly = True
+
+        self.formset = [{"-Match Values": ["_mcode", "_mnum", "_mtype", "=", "_tiebreak", "||", "_tba", "=", "_ceremonies", "_eday", "_end"],
+                         "-Status Output-": ["_output"],
+                         "Event Values-": ["_where", "=", "_prodteam", "||", "_twit", "||", "_fb", "=", "_weblink", "||", "_ename", "||", "_ecode", "=", "_pID", "||", "_tbaID", "||", "_tbaSecret", "=", "_description"]},
+                        '=', (' ', '_button', ' ')]
+
+        self._button = ControlButton('Submit')
+
+        # Add Default Text
+        self._tbaID.value += "Go to thebluealliance.com/request/apiwrite to get keys"
+        self._tbaSecret.value += "Go to thebluealliance.com/request/apiwrite to get keys"
+        self._description.value += "Add alternate description here."
+        self._mcode.value += "0"
+        self._end.value += "Only for batch uploads"
+
+        # Add ControlCombo values
+        self._where += ("Parent Folder", "../")
+        self._where += ("Same Folder", "")
+        self._mtype += ("Qualifications", "qm")
+        self._mtype += ("Quarterfinals", "qf")
+        self._mtype += ("Semifinals", "sf")
+        self._mtype += ("Finals", "f1m")
+        self._ceremonies += ("None", 0)
+        self._ceremonies += ("Opening Ceremonies", 1)
+        self._ceremonies += ("Alliance Selection", 2)
+        self._ceremonies += ("Closing Ceremonies", 3)
+        self._ceremonies += ("Highlight Reel", 4)
+        self._eday += ("Ignore", 0)
+        self._eday += ("1", 1)
+        self._eday += ("2", 2)
+        self._eday += ("3", 3)
+
+        # Define the button action
+        self._button.value = self.__buttonAction
+        self._topbutton.value = self.__buttonAction
+
+        # Get latest values from form_values.csv
         try:
             with open('form_values.csv') as csvfile:
                 reader = csv.reader(csvfile, delimiter=',', quotechar='|')
                 i = 0
                 for row in reader:
-                    for value in row:
-                        if value is not "":
+                    for val in row:
+                        if val is not "":
                             switcher = {
-                                0: myform.where,
-                                1: myform.prodteam,
-                                2: myform.twit,
-                                3: myform.fb,
-                                4: myform.weblink,
-                                5: myform.ename,
-                                6: myform.ecode,
-                                7: myform.pID,
-                                8: myform.tbaID,
-                                9: myform.tbaSecret,
-                                10: myform.description,
-                                11: myform.mcode,
-                                12: myform.mnum,
-                                13: myform.mtype,
-                                14: myform.tiebreak,
-                                15: myform.tba,
-                                16: myform.ceremonies,
-                                17: myform.eday,
-                                18: myform.end,
+                                0: self._where,
+                                1: self._prodteam,
+                                2: self._twit,
+                                3: self._fb,
+                                4: self._weblink,
+                                5: self._ename,
+                                6: self._ecode,
+                                7: self._pID,
+                                8: self._tbaID,
+                                9: self._tbaSecret,
+                                10: self._description,
+                                11: self._mcode,
+                                12: self._mnum,
+                                13: self._mtype,
+                                14: self._tiebreak,
+                                15: self._tba,
+                                16: self._ceremonies,
+                                17: self._eday,
+                                18: self._end,
                             }
-                            switcher[i].set_value(value)
+                            if any(i == k for k in (14, 15)):
+                                if val == "no":
+                                    switcher[i].value = False
+                                else:
+                                    switcher[i].value = True
+                            else:
+                                switcher[i].value = val
                         i = i + 1
                     break
         except IOError:
             print("No form_values.csv to read from, continuing with default values")
-        return render.forms(myform)
 
-    def POST(self):
-        then = datetime.now()  # tracking for the time delta
-        myform = dataform()
+    def __buttonAction(self):
+        """Button action event"""
+        then = datetime.now()
+        options = Namespace()
         row = None
-        if not myform.validates():
-            return render.forms(myform)
-        else:
-            try:
-                reader = csv.reader(open('form_values.csv'))
+        try:
+            reader = csv.reader(open('form_values.csv'))
+            row = next(reader)
+        except (StopIteration, IOError) as e:
+            with open("form_values.csv", "w+b") as csvf:  # if the file doesn't exist
+                csvf.write(''.join(str(x) for x in [","] * 20))
+                reader = csv.reader(open("form_values.csv"))
                 row = next(reader)
-            except StopIteration:
-                with open("form_values.csv", "w+b") as csvf:  # if the file doesn't exist
-                    csvf.write(''.join(str(x) for x in [","] * 30))
-                    reader = csv.reader(open("form_values.csv"))
-                    row = next(reader)
-            except IOError:
-                with open("form_values.csv", "w+b") as csvf:  # if the file doesn't exist
-                    csvf.write(''.join(str(x) for x in [","] * 30))
-                    csvf.close()
-                    reader = csv.reader(open("form_values.csv"))
-                    row = next(reader)
-            if "thebluealliance" in myform.d.mcode:
-                myform.mcode.set_value(myform.d.mcode.split("_")[-1])
-            options = Namespace()
-            formdata = web.input()
-            options.then = then
-            options.gui = True
-            options.where = row[0] = myform.d.where
-            options.prodteam = row[1] = myform.d.prodteam
-            options.twit = row[2] = myform.d.twit
-            options.fb = row[3] = myform.d.fb
-            options.weblink = row[4] = myform.d.weblink
-            options.ename = row[5] = myform.d.ename
-            options.ecode = row[6] = myform.d.ecode
-            options.pID = row[7] = myform.d.pID
-            options.tbaID = row[8] = myform.d.tbaID
-            options.tbaSecret = row[9] = myform.d.tbaSecret
-            options.description = row[10] = myform.d.description
-            options.mcode = row[11] = myform.d.mcode
-            options.mnum = row[12] = int(myform.d.mnum)
-            options.mtype = row[13] = myform.d.mtype
-            options.tiebreak = 0 if myform.d.tiebreak == "no" else 1
-            row[14] = myform.d.tiebreak
-            options.tba = 0 if myform.d.tba == "no" else 1
-            row[15] = myform.d.tba
-            options.ceremonies = row[16] = myform.d.ceremonies
-            options.eday = row[17] = myform.d.eday
-            options.end = row[18] = myform.d.end
-            thr = threading.Thread(target=yup.init, args=(options,))
-            thr.daemon = True
-            thr.start()
-            if int(myform.d.ceremonies) == 0:
-                if myform.d.end == "Only for batch uploads":
-                    myform.mnum.set_value(str(int(myform.d.mnum) + 1))
-                else:
-                    myform.mnum.set_value(str(int(myform.d.end) + 1))
-                    myform.end.set_value("Only for batch uploads")
-            elif int(myform.d.ceremonies) == 2:
-                myform.mnum.set_value("1")
-                myform.mcode.set_value("qf")
-            if myform.d.mcode == "qm" and myform.d.tiebreak == "yes":
-                myform.tiebreak.set_value("no")
-            row[12] = int(myform.d.mnum)
-            row[18] = myform.d.end
-            writer = csv.writer(open('form_values.csv', 'w'))
-            writer.writerow(row)
-            return render.forms(myform)
+        options.then = then
+        options.where = row[0] = self._where.value
+        options.prodteam = row[1] = self._prodteam.value
+        options.twit = row[2] = self._twit.value
+        options.fb = row[3] = self._fb.value
+        options.weblink = row[4] = self._weblink.value
+        options.ename = row[5] = self._ename.value
+        options.ecode = row[6] = self._ecode.value
+        options.pID = row[7] = self._pID.value
+        options.tbaID = row[8] = self._tbaID.value
+        options.tbaSecret = row[9] = self._tbaSecret.value
+        options.description = row[10] = self._description.value
+        options.mcode = row[11] = self._mcode.value
+        options.mnum = row[12] = int(self._mnum.value)
+        options.mtype = row[13] = self._mtype.value
+        options.tiebreak, row[14] = (
+            1, "yes") if self._tiebreak.value else (0, "no")
+        options.tba, row[15] = (1, "yes") if self._tba.value else (0, "no")
+        options.ceremonies = row[16] = self._ceremonies.value
+        options.eday = row[17] = self._eday.value
+        options.end = row[18] = self._end.value
+        thr = threading.Thread(target=yup.init, args=(options,))
+        thr.daemon = True
+        # thr.start()
+        if int(self._ceremonies.value) == 0:
+            if self._end.value == "Only for batch uploads":
+                self._mnum.value = str(int(self._mnum.value) + 1)
+            else:
+                self._mnum.value = str(int(self._end.value) + 1)
+                self._end.value = "Only for batch uploads"
+        elif int(self._ceremonies.value) == 2:
+            self._mnum.value = "1"
+            self._mtype.value = "qf"
+        if self._mtype.value == "qm" and self._tiebreak.value:
+            self._tiebreak.value = False
+        row[12] = int(self._mnum.value)
+        row[18] = self._end.value
+        writer = csv.writer(open('form_values.csv', 'w'))
+        writer.writerow(row)
+
+    def writePrint(self, text):
+        self._output.value += text
 
 
 def internet(host="www.google.com", port=80, timeout=4):
     try:
         host = socket.gethostbyname(host)
         socket.setdefaulttimeout(timeout)
-        socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect((host, port))
+        s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        s.connect((host, port))
+        s.close()
         return True
-    except Exception:
+    except Exception as e:
+        print(e)
         print("No internet!")
         return False
 
@@ -220,17 +220,8 @@ def main():
             subprocess.call(['sudo', 'python', sys.argv[0]])
     YA.get_youtube_service()
     YA.get_spreadsheet_service()
-    web.internalerror = web.debugerror
     if internet():
-        t = index()
-        t.daemon = True
-        t.start()
-        while True:
-            try:
-                sleep(100)
-            except KeyboardInterrupt:
-                print("\nQuitting Program")
-                return
+        pyforms.start_app(FRC_Uploader, geometry=(0, 0, 400, 550))
     else:
         return
